Add some images and basic style to recipe items in index.html
For each item representing a recipe in index.html, an enclosing
<div> has been added, whose background is equal to the image of
the final recipe. This allows the text of the name of the recipe
to be over the image itself. Now, it is more immediate to see
the corresponding dish.

To improve readibility, a basic CSS styling has been applied.
Temporarly, the style is done direcly in the .html file,
through the <style> element in the <head>.